+++ CHANGELOG.md
@@ -52,6 +52,8 @@
   * The method `onReload` has been removed as it was deprecated in v1.2.0.
   * The method `registerSearchInput` has been renamed to `registerQuickFilter`.
   * The public attribute `searchInputComponent` has been renamed to `quickFilterComponent`.
+  * The output `onListDataLoaded` has been renamed to `onDataLoaded`.
+  * The output `onPaginatedListDataLoaded` has been renamed to `onPaginatedDataLoaded`.
 * **o-table**: Attribute `show-table-buttons-text` has been renamed to `show-buttons-text`.
 * **o-form**: Output `onFormDataLoaded` has been renamed to `onDataLoaded`.
 * **o-user-info**: removed input `use-flag-icons`. This component doesn't contain the language selector anymore.